requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
eventlet==0.36.1
pymongo==4.6.1
python-dotenv==1.0.1
gunicorn==21.2.0
